requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
langchain-openai>=0.1.17
langchain-core>=0.2.24
langchain-chroma>=0.1.1
chromadb>=0.5.5
tavily-python>=0.3.5
jinja2>=3.1.4
sse-starlette>=2.1.0

